:recycle: (src) refactoring
