+++ src/apollo/index.ts
@@ -1,16 +1,16 @@
 import { AsyncStorage } from 'react-native'
-import {
-  ApolloProvider,
-  Query,
-  Mutation,
-  useQuery,
-  useMutation
-} from 'react-apollo'
+import { ApolloProvider, useQuery, useMutation } from 'react-apollo'
 import { ApolloClient } from 'apollo-client'
 import { InMemoryCache } from 'apollo-cache-inmemory'
 import { HttpLink } from 'apollo-link-http'
 import { persistCache } from 'apollo-cache-persist'
 import { resolvers, initialState } from './resolvers'
+import {
+  GET_REPO_ALL_DATA,
+  GET_FAV_ITEMS,
+  ADD_FAV_ITEM,
+  DELETE_FAV_ITEM
+} from './query'
 
 const cache = new InMemoryCache()
 // @ts-ignore
@@ -30,8 +30,10 @@ cache.writeData({ data: { ...initialState, __typename: 'favItems' } })
 export {
   client,
   ApolloProvider as Provider,
-  Query,
-  Mutation,
   useQuery,
-  useMutation
+  useMutation,
+  GET_REPO_ALL_DATA,
+  GET_FAV_ITEMS,
+  ADD_FAV_ITEM,
+  DELETE_FAV_ITEM
 }
